requirements_base.txt
appdirs==1.4.3
appnope==0.1.0
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
Babel==2.3.4
boto==2.42.0
channels==0.17.3
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
cycler==0.10.0
Cython==0.25.2
daftHM==0.0.1
daphne==0.14.3
decorator==4.0.10
dj-database-url==0.4.1
Django==1.8.8
django-countries==4.0
django-easymoney==0.7.1
django-floppyforms==1.7.0
django-idmap==0.4.1
django-vanilla-views==1.0.4
djangorestframework==3.4.6
entrypoints==0.2.2
feather-format==0.3.1
honcho==0.7.1
huey==1.2.0
IPy==0.83
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
msgpack-python==0.4.8
nbconvert==4.2.0
nbformat==4.2.0
notebook==4.3.0
numpy==1.11.2
otree-core==1.2.8
otree-save-the-change==1.1.3
packaging==16.8
pandas==0.19.2
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pyparsing==2.2.0
pytest==2.9.2
pytest-django==3.0.0
python-dateutil==2.6.0
python-redis-lock==3.2.0
pytz==2016.6.1
pyzmq==16.0.2
qtconsole==4.2.1
raven==5.25.0
redis==2.10.5
requests==2.11.1
schema==0.6.2
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
Twisted==16.2.0
txaio==2.5.1
unicodecsv==0.14.1
wcwidth==0.1.7
whitenoise==3.2.1
widgetsnbextension==1.2.6
ws4py==0.3.5
XlsxWriter==0.9.3
zope.interface==4.2.0
